refact: move toot pagination structures to core.mastodon.feed.profile.toot.pagination
