+++ core/mastodon/src/main/java/com/jeanbarrossilva/orca/core/mastodon/feed/FeedTootPaginateSource.kt
@@ -1,6 +1,6 @@
 package com.jeanbarrossilva.orca.core.mastodon.feed
 
-import com.jeanbarrossilva.orca.core.mastodon.feed.profile.toot.status.TootPaginateSource
+import com.jeanbarrossilva.orca.core.mastodon.feed.profile.toot.pagination.TootPaginateSource
 
 object FeedTootPaginateSource : TootPaginateSource() {
     override val route = "/api/v1/timelines/home"
