+++ core/mastodon/src/main/java/com/jeanbarrossilva/orca/core/mastodon/feed/profile/ProfileTootPaginateSource.kt
@@ -1,6 +1,6 @@
 package com.jeanbarrossilva.orca.core.mastodon.feed.profile
 
-import com.jeanbarrossilva.orca.core.mastodon.feed.profile.toot.status.TootPaginateSource
+import com.jeanbarrossilva.orca.core.mastodon.feed.profile.toot.pagination.TootPaginateSource
 
 class ProfileTootPaginateSource(id: String) : TootPaginateSource() {
     override val route = "/api/v1/accounts/$id/statuses"
